+++ src/Communication/SocketChannel.cpp
@@ -49,7 +49,7 @@ public:
                 continue;
             }
 
-            (void)memcpy(&_writeBuffer[_writeIndex], bufferPointer, sizeToCopy);
+            (void)memcpy(&_writeBuffer[static_cast<size_t>(_writeIndex)], bufferPointer, static_cast<size_t>(sizeToCopy));
             _writeIndex += sizeToCopy;
             bufferPointer += sizeToCopy;
             size -= sizeToCopy;
@@ -107,7 +107,7 @@ public:
                 continue;
             }
 
-            (void)memcpy(bufferPointer, &_readBuffer[_readIndex], sizeToCopy);
+            (void)memcpy(bufferPointer, &_readBuffer[static_cast<size_t>(_readIndex)], static_cast<size_t>(sizeToCopy));
             _readIndex += sizeToCopy;
             bufferPointer += sizeToCopy;
             size -= sizeToCopy;
@@ -125,7 +125,7 @@ private:
         // Did we read more than one frame the last time?
         if (_writeIndex > _endFrameIndex) {
             const int32_t bytesToMove = _writeIndex - _endFrameIndex;
-            (void)memcpy(_readBuffer.data(), &_readBuffer[_endFrameIndex], bytesToMove);
+            (void)memcpy(_readBuffer.data(), &_readBuffer[static_cast<size_t>(_endFrameIndex)], static_cast<size_t>(bytesToMove));
 
             _writeIndex -= _endFrameIndex;
 
@@ -147,7 +147,7 @@ private:
 
         while (sizeToRead > 0) {
             int32_t receivedSize{};
-            CheckResult(_socket->Receive(&_readBuffer[_writeIndex], sizeToRead, receivedSize));
+            CheckResult(_socket->Receive(&_readBuffer[static_cast<size_t>(_writeIndex)], sizeToRead, receivedSize));
 
             sizeToRead -= receivedSize;
             _writeIndex += receivedSize;
